refactor: renaming goal_processor component -> goal resolver
