--- agents/goal_processing/__init__.py
@@ -1,4 +1,0 @@
-from agents.goal_processing.base import BaseGoalResolver
-from agents.goal_processing.implicit_goal_resolver import ImplicitGoalResolver
-
-__all__ = ["BaseGoalResolver", "ImplicitGoalResolver"]
