+++ agents/goal_processor/base.py
